feat: add most of logic to games endpoints
this commit also contains some changes to how update endpoints are working and
how authorisation middleware is sharing user data throught the stack.
Previously we were sending onyl email but now we send user if available.
If not, we would fail request in it instead moving it up to handlers